backgroundScirpt refactored
- backgroundScript
+ guiBehaviour
~ variables and HUD (Hierarchy) elements refactored to more telling names
