+++ Abschlussprojekt_MMP/Assets/Game Level/Level1.unity
@@ -903,7 +903,7 @@ GameObject:
   - component: {fileID: 205833599}
   - component: {fileID: 205833598}
   m_Layer: 5
-  m_Name: score2
+  m_Name: coins
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -925,8 +925,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 1, y: 0}
   m_AnchorMax: {x: 1, y: 0}
-  m_AnchoredPosition: {x: -65, y: 681.64}
-  m_SizeDelta: {x: 45, y: 45}
+  m_AnchoredPosition: {x: -106.49, y: 658.2}
+  m_SizeDelta: {x: 64.805, y: 25.68}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &205833598
 MonoBehaviour:
@@ -1209,7 +1209,7 @@ GameObject:
   - component: {fileID: 434884943}
   - component: {fileID: 434884947}
   m_Layer: 5
-  m_Name: Canvas
+  m_Name: HUD
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -1316,7 +1316,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 3b0da1f0b6e457a4ea6896059bd43379, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
-  score1: 0
+  points_score: 0
 --- !u!1 &447339961
 GameObject:
   m_ObjectHideFlags: 0
@@ -2511,7 +2511,7 @@ Transform:
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children: []
   m_Father: {fileID: 0}
-  m_RootOrder: 4
+  m_RootOrder: 3
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &792611124
 GameObject:
@@ -2937,123 +2937,6 @@ Transform:
   m_Father: {fileID: 489434208}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
---- !u!1 &962781410
-GameObject:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  serializedVersion: 6
-  m_Component:
-  - component: {fileID: 962781414}
-  - component: {fileID: 962781413}
-  - component: {fileID: 962781412}
-  - component: {fileID: 962781411}
-  m_Layer: 0
-  m_Name: player
-  m_TagString: Untagged
-  m_Icon: {fileID: 0}
-  m_NavMeshLayer: 0
-  m_StaticEditorFlags: 0
-  m_IsActive: 0
---- !u!114 &962781411
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 962781410}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: 0c072edc2cb03e34eb886b92fe5eb997, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
---- !u!50 &962781412
-Rigidbody2D:
-  serializedVersion: 4
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 962781410}
-  m_BodyType: 0
-  m_Simulated: 1
-  m_UseFullKinematicContacts: 0
-  m_UseAutoMass: 0
-  m_Mass: 1
-  m_LinearDrag: 0
-  m_AngularDrag: 0.05
-  m_GravityScale: 1
-  m_Material: {fileID: 0}
-  m_Interpolate: 0
-  m_SleepingMode: 1
-  m_CollisionDetection: 0
-  m_Constraints: 0
---- !u!212 &962781413
-SpriteRenderer:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 962781410}
-  m_Enabled: 1
-  m_CastShadows: 0
-  m_ReceiveShadows: 0
-  m_DynamicOccludee: 1
-  m_MotionVectors: 1
-  m_LightProbeUsage: 1
-  m_ReflectionProbeUsage: 1
-  m_RayTracingMode: 0
-  m_RayTraceProcedural: 0
-  m_RenderingLayerMask: 1
-  m_RendererPriority: 0
-  m_Materials:
-  - {fileID: 10754, guid: 0000000000000000f000000000000000, type: 0}
-  m_StaticBatchInfo:
-    firstSubMesh: 0
-    subMeshCount: 0
-  m_StaticBatchRoot: {fileID: 0}
-  m_ProbeAnchor: {fileID: 0}
-  m_LightProbeVolumeOverride: {fileID: 0}
-  m_ScaleInLightmap: 1
-  m_ReceiveGI: 1
-  m_PreserveUVs: 0
-  m_IgnoreNormalsForChartDetection: 0
-  m_ImportantGI: 0
-  m_StitchLightmapSeams: 1
-  m_SelectedEditorRenderState: 0
-  m_MinimumChartSize: 4
-  m_AutoUVMaxDistance: 0.5
-  m_AutoUVMaxAngle: 89
-  m_LightmapParameters: {fileID: 0}
-  m_SortingLayerID: -1159216577
-  m_SortingLayer: 0
-  m_SortingOrder: 1
-  m_Sprite: {fileID: 21300000, guid: 2e57fc63b53a27640abc063767d330a5, type: 3}
-  m_Color: {r: 1, g: 1, b: 1, a: 1}
-  m_FlipX: 0
-  m_FlipY: 0
-  m_DrawMode: 0
-  m_Size: {x: 5, y: 5}
-  m_AdaptiveModeThreshold: 0.5
-  m_SpriteTileMode: 0
-  m_WasSpriteAssigned: 0
-  m_MaskInteraction: 0
-  m_SpriteSortPoint: 0
---- !u!4 &962781414
-Transform:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 962781410}
-  m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: -1.5258093, y: -0.19063973, z: 0}
-  m_LocalScale: {x: 0.3, y: 0.3, z: 0.3}
-  m_Children: []
-  m_Father: {fileID: 0}
-  m_RootOrder: 3
-  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &982986659
 GameObject:
   m_ObjectHideFlags: 0
@@ -3148,7 +3031,7 @@ RectTransform:
   m_Children:
   - {fileID: 1443395541}
   m_Father: {fileID: 0}
-  m_RootOrder: 5
+  m_RootOrder: 4
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 0, y: 0}
   m_AnchorMax: {x: 0, y: 0}
@@ -3168,7 +3051,7 @@ MonoBehaviour:
   m_Name: 
   m_EditorClassIdentifier: 
   speed: 3
-  bgS: {fileID: 434884947}
+  bgS: {fileID: 0}
 --- !u!1 &1022489020
 GameObject:
   m_ObjectHideFlags: 0
@@ -3621,7 +3504,7 @@ GameObject:
   - component: {fileID: 1223713509}
   - component: {fileID: 1223713508}
   m_Layer: 5
-  m_Name: healthBarBack
+  m_Name: timerBarBack
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -4212,7 +4095,7 @@ GameObject:
   - component: {fileID: 1396845725}
   - component: {fileID: 1396845724}
   m_Layer: 5
-  m_Name: score1
+  m_Name: points_score
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -4234,8 +4117,8 @@ RectTransform:
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
   m_AnchorMin: {x: 1, y: 0}
   m_AnchorMax: {x: 1, y: 0}
-  m_AnchoredPosition: {x: -110.23999, y: 681.4}
-  m_SizeDelta: {x: 45.48, y: 45.48}
+  m_AnchoredPosition: {x: -106.49, y: 686}
+  m_SizeDelta: {x: 64.8, y: 29.92}
   m_Pivot: {x: 0.5, y: 0.5}
 --- !u!114 &1396845724
 MonoBehaviour:
@@ -4532,7 +4415,7 @@ Transform:
   m_Children:
   - {fileID: 41654865}
   m_Father: {fileID: 0}
-  m_RootOrder: 6
+  m_RootOrder: 5
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1482281219
 GameObject:
@@ -4867,7 +4750,7 @@ GameObject:
   - component: {fileID: 1717277612}
   - component: {fileID: 1717277611}
   m_Layer: 5
-  m_Name: healthBarFront
+  m_Name: timerBarFront
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -5425,7 +5308,7 @@ GameObject:
   - component: {fileID: 1842757248}
   - component: {fileID: 1842757247}
   m_Layer: 5
-  m_Name: syringe1
+  m_Name: vaccine1
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -5500,7 +5383,7 @@ GameObject:
   - component: {fileID: 1949650787}
   - component: {fileID: 1949650786}
   m_Layer: 5
-  m_Name: syringe2
+  m_Name: vaccine2
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
@@ -5896,7 +5779,7 @@ GameObject:
   - component: {fileID: 2018290952}
   - component: {fileID: 2018290951}
   m_Layer: 5
-  m_Name: syringe3
+  m_Name: vaccine3
   m_TagString: Untagged
   m_Icon: {fileID: 0}
   m_NavMeshLayer: 0
